pyproject.toml
[project]
name = "my_fastapi_project"
version = "0.1.0"
description = "A brief description of your project"
requires-python = ">=3.13"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
pydantic = {extras = ["email"], version = "^2.11.7"}
psycopg2-binary = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
